requirements.txt
asgiref==3.6.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
conda==4.9.2
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
Django==4.1.4
dnspython==2.1.0
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
image==1.5.33
ipython @ file:///tmp/build/80754af9/ipython_1614615245721/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
MarkupSafe==1.1.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
ninja==1.10.2.3
numpy==1.23.4
olefile==0.46
opencv-python==4.6.0.66
pandas==1.5.1
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1614711424473/work
pkginfo==1.7.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1613849567472/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
six @ file:///tmp/build/80754af9/six_1605205327372/work
soupsieve @ file:///tmp/build/80754af9/soupsieve_1614023247537/work
split-folders==0.5.1
sqlparse==0.4.3
torch==1.8.0
torchelastic==0.2.2
torchtext==0.9.0
torchvision==0.9.0
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
